refactor(core): extract wasm code-action context parsing
