+++ crates/tsz-core/src/api/wasm/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod code_actions;
 pub(crate) mod compiler_options;
